chore: page/address 파일 mypage안으로 이전
